+++ src/models.js
@@ -1,0 +1,18 @@
+define(function() {
+	
+	return {
+		
+		session: {
+			User: {
+				name: "string",
+				company: "string"
+			}
+		}
+		
+	};
+	
+	
+	var User = {};
+	var Company = {};
+	
+});