+++ src/pages/routes.js
@@ -1,5 +1,5 @@
 define(function(require) {
-	
+
 	var menu = require("pages/v7/menu/component");
 	var map = require("pages/v7/map/component");
 	
@@ -17,15 +17,14 @@ define(function(require) {
 			component.on.pageInit = function(e) {
 				var bindings = typeof component.bindings === "function" ? component.bindings() : component.bindings;
 				e.target.bindAll(bindings);
-				if(typeof bindings['page:init'] === "function") {
-					// there really is a need to simulate page:init?
-					bindings['page:init'].apply(this, [e]);
-				}
+				// if(typeof bindings['page:init'] === "function") {
+				// 	// there really is a need to simulate page:init?
+				// 	bindings['page:init'].apply(this, [e]);
+				// }
 				return typeof pageInit === "function" ? pageInit(e) : component;
 			};
 		}
-		
-		
+
 		return { path: path, /*template: component.template,*/ component: component };
 	}