+++ src/pages/v7/menu/component.js
@@ -1,70 +1,47 @@
 define(function(require) {
 	
 	var template = require("template7!./template.html");
-	
 	var Session = require("veldoffice/Session");
 	var Framework7 = require("framework7");
 	var EM = require("veldoffice/EM");
-
+	var V7 = require("V7");
+	
+	var state, defaults = {
+		_id: "/menu", anchors: [],
+		User: { name: "&nbsp;", company: "&nbsp;" }
+	};
 	var bindings = {
-		"page:init": function(e) {
+		data() {
+			// MAJOR TODO, spagetti stuff
+			function data_getUser() {
+				var info = Session.info_;
+				if(!info || info.then) {
+					return (state && state.User) || defaults.User;
+				}
+				return ((state || defaults).User = { 
+					name: info.name, company: info.bedrijf.naam });
+			}
 			
-			// check session, hook session, keep alive?
-			// anchors...
+			var menu = V7.objects.get("/menu");
+			menu.User = data_getUser();
+			if(!menu.anchors) {
+				menu.anchors = [];
+			}
 			
-			V7.objects.get("v7-menu", function(menu) {
-				js.mixIn(menu, { 
-					anchors: [{
-						html: '<li><a href="/veldoffice/meetpunt?key=10970526" class="item-link item-content">\
-							<div class="item-media">\
-								<i class="fa fa-fw fa-street-view menu-icon bg-color-yellow"></i>\
-							</div>\
-							<div class="item-inner">\
-								<div class="item-title">\
-									<div class="item-header">Meetpunt</div>\
-									peilbuis 1\
-								</div>\
-								<div class="item-after">Bewerk</div>\
-							</div>\
-						</a></li>'
-					}, {
-						html: '<li><a href="/veldoffice/onderzoek?key=10970526" class="item-content item-link">\
-							<div class="item-media">\
-								<i class="fa fa-fw fa-street-view menu-icon bg-color-green "></i>\
-							</div>\
-							<div class="item-inner">\
-								<div class="item-title">\
-									<div class="item-header"><div class="" style="font-weight:bold;">2e Nieuwstraat 43-47 te Hilversum</div></div>\
-									<div class="item-text" style="max-height:none;">\
-										<div>\
-											<span class="chip methode">NEN 5104</span>\
-											<div class="chip LOPEND"><div class="chip-label">In uitvoering</div></div>\
-											<div class="chip show-on-map"><div class="chip-label">Kaart</div></div>\
-										</div>\
-										<div class="item-summary">\
-											<span title="27 foto\'s">\
-												<i class="fa fa-camera"></i> 27\
-											</span>\
-											<span title="11 meetpunten" class="">\
-												<i class="fa fa-dot-circle-o"></i> 9 / 11\
-												<span style="color:silver; font-size:9pt;"> - Hopman en Peters</span>\
-											</span>\
-										</div>\
-									</div>\
-								</div>\
-							</div>\
-						</a></li>'
-					}]
-				});
-			});
+			V7.objects.save(menu);
+			
+			return menu;
 		},
+
 		".page-content infinite": function() {
 			console.log("infinite", { 'this': this, args: arguments });
 		},
 		".ptr-content ptr:refresh": function(e) {
+			state = undefined;
 			Session.refresh().then(function() {
 				f7a.ptr.done(e.target);
 				V7.sessionNeeded();
+				V7.refreshMenu(); //?
 			});
 		},
 		".link.logout click": function() {
@@ -116,59 +93,8 @@ define(function(require) {
 			});
 		}
 	};
-	
-	function data() {
-		var info = Session.info_;
-		
-		if(!info || info.then) {
-			info = { name: "", bedrijf: { naam: "" } };
-		}
-		
-		return {
-			User: { name: info.name, company: info.bedrijf.naam },
-			anchors: [{
-				html: '<li><a href="/veldoffice/meetpunt?key=10970526" class="item-link item-content">\
-					<div class="item-media">\
-						<i class="fa fa-fw fa-street-view menu-icon bg-color-yellow"></i>\
-					</div>\
-					<div class="item-inner">\
-						<div class="item-title">\
-							<div class="item-header">Meetpunt</div>\
-							peilbuis 1\
-						</div>\
-						<div class="item-after">Bewerk</div>\
-					</div>\
-				</a></li>'
-			}, {
-				html: '<li><a href="/veldoffice/onderzoek?key=10970526" class="item-content item-link">\
-					<div class="item-media">\
-						<i class="fa fa-fw fa-street-view menu-icon bg-color-green "></i>\
-					</div>\
-					<div class="item-inner">\
-						<div class="item-title">\
-							<div class="item-header"><div class="" style="font-weight:bold;">2e Nieuwstraat 43-47 te Hilversum</div></div>\
-							<div class="item-text" style="max-height:none;">\
-								<div>\
-									<span class="chip methode">NEN 5104</span>\
-									<div class="chip LOPEND"><div class="chip-label">In uitvoering</div></div>\
-									<div class="chip show-on-map"><div class="chip-label">Kaart</div></div>\
-								</div>\
-								<div class="item-summary">\
-									<span title="27 foto\'s">\
-										<i class="fa fa-camera"></i> 27\
-									</span>\
-									<span title="11 meetpunten" class="">\
-										<i class="fa fa-dot-circle-o"></i> 9 / 11\
-										<span style="color:silver; font-size:9pt;"> - Hopman en Peters</span>\
-									</span>\
-								</div>\
-							</div>\
-						</div>\
-					</div>\
-				</a></li>'
-			}]
-		};
-	}
 
-	return { template: template, data: data, bindings: bindings };
+	return { template: template, data: bindings.data, bindings: bindings };
 });
+
+