+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+### 2018-09-08
+- Developing anchors
+- Developing server-paged-pouchdb-cached-infinite-virtual-scrolling-list-query 
+- Introducing V7.objects and veldoffice/EM.query{pager}
+
+### 2018-09-02
+- `routes.js`
+
 ### 2018-08-22
 - Finally getting up to speed _with myf'nself_